Refactor as mutable grid and split into 'hooks'
Learned more about Vue and that it is not made to deal with immutable
data. Decided to give it a try to rewrite in a mutable style and
splitting the different logic groups into 'hooks'-style (not sure what
Vue's terminology is here).

Overall pretty pleased with the result, feels more composable :)